feat: split tests to their own test files
