+++ pyproject.toml
@@ -1,3 +1,4 @@
+# code formatter settings:
 [tool.black]
 line-length = 100
 target-version = ['py38']
@@ -18,3 +19,9 @@ exclude = '''
   | venv
 )/
 '''
+
+# Set root for pytest: https://stackoverflow.com/a/50610630
+[tool.pytest.ini_options]
+pythonpath = [
+  "."
+]
