+++ tests/mainnet/metaregistry/api/test_get_pool_asset_type.py
@@ -1,0 +1,21 @@
+@pytest.mark.parametrize("pool_id", range(MAX_POOLS))
+def test_get_pool_asset_type(metaregistry, registry_pool_index_iterator, pool_id):
+
+    skip_if_pool_id_gte_max_pools_in_registry(pool_id, registry_pool_index_iterator)
+
+    registry_id, registry_handler, registry, pool = registry_pool_index_iterator[pool_id]
+
+    if check_pool_already_registered(metaregistry, pool, registry_handler):
+        pytest.skip()
+
+    # get_pool_asset_type
+    if registry_id in [
+        METAREGISTRY_CRYPTO_REGISTRY_HANDLER_INDEX,
+        METAREGISTRY_CRYPTO_FACTORY_HANDLER_INDEX,
+    ]:
+        actual_output = 4
+    else:
+        actual_output = registry.get_pool_asset_type(pool)
+
+    metaregistry_output = metaregistry.get_pool_asset_type(pool)
+    assert actual_output == metaregistry_output
