+++ tests/mainnet/metaregistry/api/test_pool_is_registered.py
@@ -1,0 +1,2 @@
+def test_all_pools(populated_metaregistry, pool):
+    assert populated_metaregistry.is_registered(pool)
