+++ tests/mainnet/metaregistry/test_base_registry_tracking.py
@@ -1,22 +1,18 @@
 import ape
 
-from tests.utils.constants import (
-    BVECVX_LPTOKEN,
-    CVXFXS_LPTOKEN,
-    DAI,
-    METAREGISTRY_CRYPTO_FACTORY_HANDLER_INDEX,
-    METAREGISTRY_STABLE_FACTORY_HANDLER_INDEX,
-)
 
+def test_new_crypto_factory_pool(metaregistry, crypto_factory, tokens, lp_tokens, alice):
+    test_pool_name = "test_pool"
 
-def test_new_crypto_factory_pool(metaregistry, registries, owner):
-    test_pool_name = "test"
-    crypto_factory = registries[METAREGISTRY_CRYPTO_FACTORY_HANDLER_INDEX]
-    assert metaregistry.find_pool_for_coins(DAI, CVXFXS_LPTOKEN) == ape.ZERO_ADDRESS
+    assert (
+        metaregistry.find_pool_for_coins(tokens["dai"], lp_tokens["cvxFXSFXS-f"])
+        == ape.utils.ZERO_ADDRESS
+    )
+
     crypto_factory.deploy_pool(
         test_pool_name,
         test_pool_name,
-        [DAI, CVXFXS_LPTOKEN],
+        [tokens["dai"], lp_tokens["cvxFXSFXS-f"]],
         200000000,
         100000000000000,
         1000000,
@@ -27,36 +23,54 @@ def test_new_crypto_factory_pool(metaregistry, registries, owner):
         5000000000,
         600,
         994000214704046300,
-        {"from": owner},
+        sender=alice,
     )
     new_pool = crypto_factory.pool_list(crypto_factory.pool_count() - 1)
     lp_token = crypto_factory.get_token(new_pool)
-    assert metaregistry.get_coins(new_pool) == [DAI, CVXFXS_LPTOKEN] + [ape.ZERO_ADDRESS] * 6
+
+    assert (
+        metaregistry.get_coins(new_pool)
+        == [tokens["dai"], lp_tokens["cvxFXSFXS-f"]] + [ape.utils.ZERO_ADDRESS] * 6
+    )
     assert test_pool_name in metaregistry.get_pool_name(new_pool)
     assert metaregistry.get_pool_from_lp_token(lp_token) == new_pool
-    assert metaregistry.find_pool_for_coins(DAI, CVXFXS_LPTOKEN) == new_pool
+    assert metaregistry.find_pool_for_coins(tokens["dai"], lp_tokens["cvxFXSFXS-f"]) == new_pool
 
 
-def test_new_stable_factory_pool(metaregistry, registries, owner):
-    test_pool_name = "test2"
-    stable_factory = registries[METAREGISTRY_STABLE_FACTORY_HANDLER_INDEX]
-    assert metaregistry.find_pool_for_coins(BVECVX_LPTOKEN, CVXFXS_LPTOKEN) == ape.ZERO_ADDRESS
+def test_new_stable_factory_pool(metaregistry, stable_factory, lp_tokens, alice):
+    test_pool_name = "test_pool"
+
+    assert (
+        metaregistry.find_pool_for_coins(lp_tokens["bveCVX-CVX-f"], lp_tokens["cvxFXSFXS-f"])
+        == ape.utils.ZERO_ADDRESS
+    )
+
     stable_factory.deploy_plain_pool(
         test_pool_name,
         test_pool_name,
-        [BVECVX_LPTOKEN, CVXFXS_LPTOKEN, ape.ZERO_ADDRESS, ape.ZERO_ADDRESS],
+        [
+            lp_tokens["bveCVX-CVX-f"],
+            lp_tokens["cvxFXSFXS-f"],
+            ape.utils.ZERO_ADDRESS,
+            ape.utils.ZERO_ADDRESS,
+        ],
         10000,
         4000000,
         0,
         0,
-        {"from": owner},
+        sender=alice,
     )
+
     new_pool = stable_factory.pool_list(stable_factory.pool_count() - 1)
-    lp_token = new_pool
+    lp_token = new_pool  # pool == lp_token fot stable_factory
+
     assert (
         metaregistry.get_coins(new_pool)
-        == [BVECVX_LPTOKEN, CVXFXS_LPTOKEN] + [ape.ZERO_ADDRESS] * 6
+        == [lp_tokens["bveCVX-CVX-f"], lp_tokens["cvxFXSFXS-f"]] + [ape.utils.ZERO_ADDRESS] * 6
     )
     assert test_pool_name in metaregistry.get_pool_name(new_pool)
     assert metaregistry.get_pool_from_lp_token(lp_token) == new_pool
-    assert metaregistry.find_pool_for_coins(BVECVX_LPTOKEN, CVXFXS_LPTOKEN) == new_pool
+    assert (
+        metaregistry.find_pool_for_coins(lp_tokens["bveCVX-CVX-f"], lp_tokens["cvxFXSFXS-f"])
+        == new_pool
+    )
