+++ tests/mainnet/registries/test_add_remove_pool.py
@@ -22,7 +22,7 @@ def test_revert_unauthorised_add_pool(crypto_registry_v1, unauthorised_account):
             TRICRYPTO2_ZAP,
             3,
             "tricrypto2",
-            ape.ZERO_ADDRESS,
+            ape.utils.ZERO_ADDRESS,
             False,
             {"from": unauthorised_account},
         )
@@ -38,7 +38,7 @@ def test_revert_add_existing_pool(crypto_registry_updated, owner):
             TRICRYPTO2_ZAP,
             3,
             "tricrypto2",
-            ape.ZERO_ADDRESS,
+            ape.utils.ZERO_ADDRESS,
             False,
             {"from": owner},
         )
@@ -56,7 +56,7 @@ def test_add_pool(crypto_registry_v1, owner):
         TRICRYPTO2_ZAP,
         3,
         "tricrypto2",
-        ape.ZERO_ADDRESS,
+        ape.utils.ZERO_ADDRESS,
         False,
         {"from": owner},
     )
@@ -66,7 +66,7 @@ def test_add_pool(crypto_registry_v1, owner):
     assert crypto_registry_v1.get_zap(TRICRYPTO2_MAINNET) == TRICRYPTO2_ZAP
     assert crypto_registry_v1.get_lp_token(TRICRYPTO2_MAINNET) == CRV3CRYPTO_MAINNET
     assert crypto_registry_v1.get_pool_from_lp_token(CRV3CRYPTO_MAINNET) == TRICRYPTO2_MAINNET
-    assert crypto_registry_v1.get_base_pool(TRICRYPTO2_MAINNET) == ape.ZERO_ADDRESS
+    assert crypto_registry_v1.get_base_pool(TRICRYPTO2_MAINNET) == ape.utils.ZERO_ADDRESS
     assert not crypto_registry_v1.is_meta(TRICRYPTO2_MAINNET)
     assert crypto_registry_v1.get_pool_name(TRICRYPTO2_MAINNET) == "tricrypto2"
 
@@ -76,7 +76,7 @@ def test_add_pool(crypto_registry_v1, owner):
     assert crypto_registry_v1.get_gauges(TRICRYPTO2_MAINNET)[1][0] == 5  # gauge type is 5
     assert (
         crypto_registry_v1.get_coins(TRICRYPTO2_MAINNET)
-        == [USDT, WBTC, WETH] + [ape.ZERO_ADDRESS] * 5
+        == [USDT, WBTC, WETH] + [ape.utils.ZERO_ADDRESS] * 5
     )
 
     # check if coins and underlying coins (if any) are added to the underlying market:
@@ -101,7 +101,7 @@ def test_remove_pool(crypto_registry_v1, owner):
         TRICRYPTO2_ZAP,
         3,
         "tricrypto2",
-        ape.ZERO_ADDRESS,
+        ape.utils.ZERO_ADDRESS,
         False,
         {"from": owner},
     )
@@ -112,10 +112,10 @@ def test_remove_pool(crypto_registry_v1, owner):
         "0x98a7F18d4E56Cfe84E3D081B40001B3d5bD3eB8B",  # _pool
         "0x3D229E1B4faab62F621eF2F6A610961f7BD7b23B",  # _lp_token
         "0x65CA7Dc5CB661fC58De57B1E1aF404649a27AD35",  # _gauge
-        ape.ZERO_ADDRESS,  # _zap
+        ape.utils.ZERO_ADDRESS,  # _zap
         2,  # _n_coins
         "eursusd",  # _name
-        ape.ZERO_ADDRESS,  # _base_pool
+        ape.utils.ZERO_ADDRESS,  # _base_pool
         False,  # _has_positive_rebasing_tokens
         {"from": owner},
     )
@@ -128,17 +128,17 @@ def test_remove_pool(crypto_registry_v1, owner):
 
     assert crypto_registry_v1.pool_list(0) == last_pool
     assert crypto_registry_v1.pool_count() == pool_count - 1  # one pool should be gone
-    assert crypto_registry_v1.get_zap(TRICRYPTO2_MAINNET) == ape.ZERO_ADDRESS
-    assert crypto_registry_v1.get_lp_token(TRICRYPTO2_MAINNET) == ape.ZERO_ADDRESS
-    assert crypto_registry_v1.get_pool_from_lp_token(CRV3CRYPTO_MAINNET) == ape.ZERO_ADDRESS
+    assert crypto_registry_v1.get_zap(TRICRYPTO2_MAINNET) == ape.utils.ZERO_ADDRESS
+    assert crypto_registry_v1.get_lp_token(TRICRYPTO2_MAINNET) == ape.utils.ZERO_ADDRESS
+    assert crypto_registry_v1.get_pool_from_lp_token(CRV3CRYPTO_MAINNET) == ape.utils.ZERO_ADDRESS
     assert crypto_registry_v1.get_pool_name(TRICRYPTO2_MAINNET) == ""
 
     assert crypto_registry_v1.get_n_coins(TRICRYPTO2_MAINNET) == 0
     assert crypto_registry_v1.get_decimals(TRICRYPTO2_MAINNET) == [0] * MAX_COINS
-    assert crypto_registry_v1.get_gauges(TRICRYPTO2_MAINNET)[0][0] == ape.ZERO_ADDRESS
+    assert crypto_registry_v1.get_gauges(TRICRYPTO2_MAINNET)[0][0] == ape.utils.ZERO_ADDRESS
     assert crypto_registry_v1.get_gauges(TRICRYPTO2_MAINNET)[1][0] == 0
 
-    assert crypto_registry_v1.get_coins(TRICRYPTO2_MAINNET) == [ape.ZERO_ADDRESS] * MAX_COINS
+    assert crypto_registry_v1.get_coins(TRICRYPTO2_MAINNET) == [ape.utils.ZERO_ADDRESS] * MAX_COINS
 
     for coin_i in [USDT, WBTC, WETH]:
         for coin_j in [USDT, WBTC, WETH]:
